Previous canvas component was fine.
